harness: dissolve state boundary into substrate
