+++ spark/harness/substrate.py
@@ -13,10 +13,14 @@ value (chat/create roles by default, off for code).
 from __future__ import annotations
 
 import asyncio
+import datetime as _dt
 import json
 import os
+import re
 import subprocess
 import sys
+import time
+import uuid
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
@@ -49,6 +53,597 @@ def load_file(path: str | os.PathLike) -> str | None:
         return None
     return content if content else None
 
+# ---------------------------------------------------------------------------
+# Session, recall, and live-state substrate
+# ---------------------------------------------------------------------------
+
+SESSIONS_DIR = Path(os.path.expanduser("~/.cache/vybn-spark/sessions"))
+FRESH_WINDOW_SEC = 24 * 3600  # 24h default
+
+
+@dataclass
+class SessionInfo:
+    session_id: str
+    path: Path
+    mtime: float
+    turn_count: int
+    preview: str  # first user turn, truncated
+
+
+class SessionStore:
+    def __init__(self, root: Path = SESSIONS_DIR) -> None:
+        self.root = root
+        self.root.mkdir(parents=True, exist_ok=True)
+        self._current_id: str | None = None
+        self._current_path: Path | None = None
+        self._last_saved_len: int = 0
+
+    # ------------------------------------------------------------------
+    # Session lifecycle
+    # ------------------------------------------------------------------
+
+    def new_session(self) -> str:
+        """Create a new session id and path. Does not write anything yet."""
+        ts = datetime.now(timezone.utc).strftime("%Y-%m-%d_%H-%M-%S")
+        sid = f"{ts}_{uuid.uuid4().hex[:8]}"
+        self._current_id = sid
+        self._current_path = self.root / f"{sid}.jsonl"
+        self._last_saved_len = 0
+        return sid
+
+    def adopt_session(self, session_id: str) -> bool:
+        """Adopt an existing session id as the current one."""
+        path = self.root / f"{session_id}.jsonl"
+        if not path.exists():
+            return False
+        self._current_id = session_id
+        self._current_path = path
+        # count how many messages are already persisted
+        self._last_saved_len = sum(1 for _ in path.open("r", encoding="utf-8"))
+        return True
+
+    @property
+    def current_id(self) -> str | None:
+        return self._current_id
+
+    # ------------------------------------------------------------------
+    # Persistence
+    # ------------------------------------------------------------------
+
+    def append_new(self, messages: list[dict]) -> int:
+        """Append any messages beyond what is already persisted.
+
+        Returns the number of messages written this call. Idempotent in the
+        sense that calling it twice with the same `messages` only writes the
+        delta once.
+        """
+        if self._current_path is None:
+            self.new_session()
+        assert self._current_path is not None
+
+        n_total = len(messages)
+        if n_total <= self._last_saved_len:
+            return 0
+
+        ts = datetime.now(timezone.utc).isoformat()
+        with self._current_path.open("a", encoding="utf-8") as f:
+            for msg in messages[self._last_saved_len:]:
+                f.write(json.dumps({"ts": ts, "msg": msg}, ensure_ascii=False) + "\n")
+
+        written = n_total - self._last_saved_len
+        self._last_saved_len = n_total
+        return written
+
+    def load(self, session_id: str) -> list[dict]:
+        """Load the messages list from a session. Returns [] if not found."""
+        path = self.root / f"{session_id}.jsonl"
+        if not path.exists():
+            return []
+        messages: list[dict] = []
+        with path.open("r", encoding="utf-8") as f:
+            for line in f:
+                line = line.strip()
+                if not line:
+                    continue
+                try:
+                    entry = json.loads(line)
+                    if isinstance(entry, dict) and "msg" in entry:
+                        messages.append(entry["msg"])
+                except json.JSONDecodeError:
+                    # skip corrupted lines (partial write on crash)
+                    continue
+        return messages
+
+    # ------------------------------------------------------------------
+    # Listing / discovery
+    # ------------------------------------------------------------------
+
+    def list_sessions(self, limit: int = 10) -> list[SessionInfo]:
+        """List recent sessions, most recent first."""
+        out: list[SessionInfo] = []
+        for p in sorted(self.root.glob("*.jsonl"), key=lambda x: x.stat().st_mtime, reverse=True):
+            sid = p.stem
+            mtime = p.stat().st_mtime
+            turn_count = 0
+            preview = ""
+            try:
+                with p.open("r", encoding="utf-8") as f:
+                    for line in f:
+                        line = line.strip()
+                        if not line:
+                            continue
+                        try:
+                            entry = json.loads(line)
+                            msg = entry.get("msg", {})
+                            if isinstance(msg, dict):
+                                turn_count += 1
+                                if not preview and msg.get("role") == "user":
+                                    content = msg.get("content", "")
+                                    if isinstance(content, str):
+                                        preview = content[:80].replace("\n", " ")
+                                    elif isinstance(content, list):
+                                        for item in content:
+                                            if isinstance(item, dict) and item.get("type") == "text":
+                                                preview = item.get("text", "")[:80].replace("\n", " ")
+                                                break
+                        except json.JSONDecodeError:
+                            continue
+            except Exception:
+                pass
+            out.append(SessionInfo(session_id=sid, path=p, mtime=mtime,
+                                    turn_count=turn_count, preview=preview or "(empty)"))
+            if len(out) >= limit:
+                break
+        return out
+
+    def latest_fresh(self, window_sec: int = FRESH_WINDOW_SEC) -> SessionInfo | None:
+        """Return the most recent session within the freshness window, or None."""
+        sessions = self.list_sessions(limit=1)
+        if not sessions:
+            return None
+        s = sessions[0]
+        if time.time() - s.mtime > window_sec:
+            return None
+        if s.turn_count == 0:
+            return None
+        return s
+
+    def format_age(self, mtime: float) -> str:
+        delta = time.time() - mtime
+        if delta < 60:
+            return f"{int(delta)}s ago"
+        if delta < 3600:
+            return f"{int(delta // 60)}m ago"
+        if delta < 86400:
+            return f"{int(delta // 3600)}h ago"
+        return f"{int(delta // 86400)}d ago"
+
+# === RECALL GATE ==========================================================
+#
+# "Read bytes before describing" applied to conversational memory. When the
+# user asks about prior conversation state ("do you recall...", "what did
+# we say about..."), the honest first move is to read the session log, not
+# to reconstruct from in-context fragments. SessionStore already owns the
+# read over ~/.cache/vybn-spark/sessions/*.jsonl; this section adds the
+# classifier + keyword probe on top of it so the agent loop can inject the
+# retrieved bytes into the live prompt layer before the model generates.
+#
+# Same move as On Describing Internals (vybn-os SKILL.md) applied to a
+# second surface. The absorb_gate binds refactor-first in the loop; this
+# binds read-session-logs in the loop.
+
+
+# Phrases that strongly indicate the user is asking about prior
+# conversation state. Tuned to fire on recall questions and stay quiet on
+# hypothetical or forward-looking ones.
+_RECALL_PATTERNS: tuple = (
+    re.compile(r"\b(do|did) you (recall|remember|recollect)\b", re.I),
+    re.compile(r"\b(you|we) (said|wrote|mentioned|talked about|discussed)\b", re.I),
+    re.compile(r"\b(earlier|before|previously|yesterday|this (morning|afternoon|evening|session))\b.*\b(say|said|write|wrote|talk|mention|discuss|bring up)", re.I),
+    re.compile(r"\bwhere (did|does|do) (our|this|the) (conversation|thread|session|chat) (begin|start)", re.I),
+    re.compile(r"\bwhat (did|does|do) (i|you|we) (say|said|write|wrote|mean|bring up)\b", re.I),
+    re.compile(r"\bremind me (what|when|where|how)\b", re.I),
+    re.compile(r"\b(the|that) thing (you|we|i) (said|mentioned|talked about|brought up)\b", re.I),
+    re.compile(r"\b(go|went) back to (what|when|where)\b", re.I),
+)
+
+_RECALL_STOPWORDS = {
+    "the", "a", "an", "and", "or", "but", "if", "then", "you", "i", "we",
+    "us", "our", "my", "your", "do", "did", "does", "recall", "remember",
+    "recollect", "said", "wrote", "say", "write", "mentioned", "talked",
+    "discussed", "talk", "discuss", "where", "when", "what", "how", "why",
+    "this", "that", "it", "is", "was", "were", "be", "been", "being",
+    "to", "of", "in", "on", "at", "for", "with", "about", "really",
+    "begin", "start", "began", "started", "go", "went", "back",
+    "remind", "me", "thing", "things", "conversation", "thread", "session",
+    "chat", "earlier", "before", "previously", "yesterday", "morning",
+    "afternoon", "evening", "tonight", "today", "bring", "brought", "up",
+    "mean", "means", "meant", "from", "than", "there", "here",
+}
+
+
+@dataclass
+class RecallHit:
+    ts: str
+    role: str
+    content: str
+    session_file: str
+
+
+def is_recall_question(text: str) -> bool:
+    """True when the message asks about prior conversation state."""
+    if not text or len(text) > 4000:
+        return False
+    return any(pat.search(text) for pat in _RECALL_PATTERNS)
+
+
+def _recall_keywords(text: str, *, max_keywords: int = 8) -> list[str]:
+    words = re.findall(r"\b[a-zA-Z][a-zA-Z'-]{2,}\b", text)
+    seen: set[str] = set()
+    out: list[str] = []
+    for w in words:
+        lw = w.lower()
+        if lw in _RECALL_STOPWORDS or lw in seen:
+            continue
+        seen.add(lw)
+        out.append(w)
+        if len(out) >= max_keywords:
+            break
+    return out
+
+
+def _recall_iter(files):
+    for f in files:
+        try:
+            with f.open() as fh:
+                for line in fh:
+                    line = line.strip()
+                    if not line:
+                        continue
+                    try:
+                        rec = json.loads(line)
+                    except json.JSONDecodeError:
+                        continue
+                    msg = rec.get("msg") or {}
+                    content = msg.get("content") or ""
+                    if not isinstance(content, str) or not content.strip():
+                        continue
+                    yield RecallHit(
+                        ts=rec.get("ts", ""),
+                        role=msg.get("role", "?"),
+                        content=content,
+                        session_file=f.name,
+                    )
+        except OSError:
+            continue
+
+
+def recent_files(hours: float) -> list:
+    if not SESSIONS_DIR.exists():
+        return []
+    cutoff = time.time() - hours * 3600
+    files = []
+    for p in SESSIONS_DIR.glob("*.jsonl"):
+        try:
+            if p.stat().st_mtime >= cutoff:
+                files.append(p)
+        except OSError:
+            continue
+    files.sort(key=lambda p: p.stat().st_mtime, reverse=True)
+    return files
+
+
+def _recall_origin_hits(hours: float, max_hits: int) -> list:
+    """Return the EARLIEST user messages from recent sessions.
+
+    Fallback for recall questions whose keywords are all stopwords
+    ("where did we begin", "what did you say earlier"). The user is
+    asking about origin/thread shape; deliver the opening turns so the
+    model can answer from them.
+    """
+    files = recent_files(hours)
+    if not files:
+        return []
+    # newest session first, but within each session take EARLIEST messages
+    out: list = []
+    seen_sig: set = set()
+    for f in files:
+        session_hits: list = []
+        try:
+            with f.open() as fh:
+                for line in fh:
+                    line = line.strip()
+                    if not line:
+                        continue
+                    try:
+                        rec = json.loads(line)
+                    except json.JSONDecodeError:
+                        continue
+                    msg = rec.get("msg") or {}
+                    if msg.get("role") != "user":
+                        continue
+                    content = msg.get("content") or ""
+                    if not isinstance(content, str) or not content.strip():
+                        continue
+                    sig = content[:200]
+                    if sig in seen_sig:
+                        continue
+                    seen_sig.add(sig)
+                    session_hits.append(RecallHit(
+                        ts=rec.get("ts", ""),
+                        role="user",
+                        content=content,
+                        session_file=f.name,
+                    ))
+                    if len(session_hits) >= 2:
+                        break  # first two user turns per session is enough
+        except OSError:
+            continue
+        out.extend(session_hits)
+        if len(out) >= max_hits:
+            break
+    return out[:max_hits]
+
+
+def search_sessions(query: str, *, hours: float = 24.0, max_hits: int = 6) -> list[RecallHit]:
+    """Recent session messages matching query keywords, ranked by hit count.
+
+    When keyword extraction yields nothing (stopword-heavy recall
+    questions like "where did our conversation begin"), fall back to the
+    earliest user messages from recent sessions — the opening turns are
+    what "where did it begin" is literally asking for.
+    """
+    keywords = _recall_keywords(query)
+    if not keywords:
+        return _recall_origin_hits(hours, max_hits)
+    if not SESSIONS_DIR.exists():
+        return []
+    files = recent_files(hours)
+    if not files:
+        return []
+    keyword_res = [re.compile(r"\b" + re.escape(k) + r"\b", re.I) for k in keywords]
+
+    scored: list = []
+    seen_sig: set = set()
+    for hit in _recall_iter(files):
+        sig = hit.content[:200]
+        if sig in seen_sig:
+            continue
+        score = sum(1 for pat in keyword_res if pat.search(hit.content))
+        if score >= 1:
+            seen_sig.add(sig)
+            scored.append((score, hit))
+    scored.sort(key=lambda t: t[0], reverse=True)
+    return [h for _, h in scored[:max_hits]]
+
+
+def format_recall_injection(hits: list, *, max_chars_per_hit: int = 800) -> str:
+    """Render hits as an explicit retrieval block for the live prompt layer."""
+    if not hits:
+        return ""
+    lines = [
+        "[recall-gate retrieval — session log bytes, not inferred]",
+        f"Query matched {len(hits)} message(s) from prior session logs.",
+        "Answer from these bytes. If the answer isn't here, say so — do",
+        "not reconstruct from pattern.",
+        "",
+    ]
+    for i, h in enumerate(hits, 1):
+        content = h.content
+        if len(content) > max_chars_per_hit:
+            content = content[:max_chars_per_hit] + "…[truncated]"
+        lines.append(f"--- hit {i} [{h.role} @ {h.ts}] ({h.session_file}) ---")
+        lines.append(content)
+        lines.append("")
+    return "\n".join(lines).rstrip() + "\n"
+
+
+def maybe_recall_probe(user_text: str, *, hours: float = 24.0) -> tuple[bool, str, int]:
+    """Single entry point for the agent loop.
+
+    Returns (triggered, injection_text, hit_count). When triggered is True
+    and hit_count is 0, a short note still returns so the model names the
+    retrieval gap instead of confabulating from silence.
+    """
+    if not is_recall_question(user_text):
+        return False, "", 0
+    hits = search_sessions(user_text, hours=hours)
+    if not hits:
+        note = (
+            "[recall-gate retrieval — session log bytes, not inferred]\n"
+            f"Query was classified as a recall question but produced zero\n"
+            f"matches in the last {hours:.0f} hours of session logs. Answer\n"
+            "by naming the retrieval gap, not by reconstructing from pattern.\n"
+        )
+        return True, note, 0
+    return True, format_recall_injection(hits), len(hits)
+
+
+# === LIVE SNAPSHOT ========================================================
+
+_REPOS = [
+    ("Vybn",       "~/Vybn",       "main"),
+    ("Him",        "~/Him",        "main"),
+    ("Vybn-Law",   "~/Vybn-Law",   "master"),
+    ("vybn-phase", "~/vybn-phase", "main"),
+]
+
+_GH_REPO = "zoedolan/Vybn"
+
+
+def _run(cmd: list[str], *, cwd: Optional[str] = None, timeout: float = 6.0) -> str:
+    """Run a command, return stripped stdout, swallow everything else.
+
+    list[str] invocation (no shell=True) keeps this safe against exotic
+    repo paths and avoids any accidental shell interpretation.
+    """
+    try:
+        r = subprocess.run(
+            cmd,
+            cwd=cwd,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.DEVNULL,
+            text=True,
+            timeout=timeout,
+        )
+        return (r.stdout or "").strip()
+    except Exception:
+        return ""
+
+
+def _expand(path: str) -> str:
+    return os.path.expanduser(path)
+
+
+def _repo_block(name: str, path: str, branch: str, *, timeout: float) -> str:
+    exp = _expand(path)
+    if not Path(exp).exists():
+        return f"{name}: (not checked out at {path})"
+
+    head_short = _run(["git", "rev-parse", "--short", "HEAD"], cwd=exp, timeout=timeout) or "?"
+    cur_branch = _run(["git", "rev-parse", "--abbrev-ref", "HEAD"], cwd=exp, timeout=timeout) or branch
+    log = _run(["git", "log", "--oneline", "-5"], cwd=exp, timeout=timeout)
+    status = _run(["git", "status", "--short"], cwd=exp, timeout=timeout)
+    ahead_behind = _run(
+        ["git", "rev-list", "--left-right", "--count", f"origin/{branch}...HEAD"],
+        cwd=exp,
+        timeout=timeout,
+    )
+
+    if status:
+        dirty_count = len([ln for ln in status.splitlines() if ln.strip()])
+        dirty_note = f"{dirty_count} uncommitted"
+    else:
+        dirty_note = "clean"
+
+    ab_note = ""
+    if ahead_behind and "\t" in ahead_behind:
+        parts = ahead_behind.split()
+        if len(parts) == 2:
+            behind, ahead = parts
+            if ahead != "0" or behind != "0":
+                ab_note = f", {ahead} ahead / {behind} behind origin/{branch}"
+
+    lines = [f"{name} [{cur_branch} @ {head_short}] — {dirty_note}{ab_note}"]
+    if log:
+        for ln in log.splitlines():
+            lines.append(f"  {ln}")
+    else:
+        lines.append("  (no git log)")
+    return "\n".join(lines)
+
+
+def _pr_block(*, timeout: float) -> tuple[str, int | None]:
+    """Return (formatted_block, highest_pr_number_or_None)."""
+    out = _run(
+        [
+            "gh", "pr", "list",
+            "--state", "all",
+            "--limit", "15",
+            "--json", "number,title,state,headRefName",
+            "--repo", _GH_REPO,
+        ],
+        timeout=timeout,
+    )
+    if not out:
+        return ("(gh pr list unavailable — offline or rate-limited)", None)
+    try:
+        prs = json.loads(out)
+    except Exception:
+        return ("(gh pr list returned unparseable JSON)", None)
+    if not prs:
+        return ("(no recent PRs)", None)
+    lines = []
+    highest = None
+    for pr in prs[:10]:
+        state = str(pr.get("state", "?")).upper()[:6]
+        num = pr.get("number")
+        title = str(pr.get("title", "?"))[:72]
+        branch = str(pr.get("headRefName", "?"))[:32]
+        if isinstance(num, int):
+            if highest is None or num > highest:
+                highest = num
+            lines.append(f"  #{num} [{state:<6}] {title}  ({branch})")
+    return ("\n".join(lines) if lines else "(no PRs)", highest)
+
+
+_PR_REF_RE = re.compile(r"\bPR\s*#\s*(\d+)", re.IGNORECASE)
+
+
+def _continuity_drift(continuity_path: str, current_pr: int | None) -> str:
+    p = Path(_expand(continuity_path))
+    if not p.exists():
+        return ""
+    try:
+        text = p.read_text(encoding="utf-8", errors="replace")
+    except OSError:
+        return ""
+    refs = [int(m.group(1)) for m in _PR_REF_RE.finditer(text)]
+    if not refs:
+        return ""
+    last_cont = max(refs)
+    if current_pr is None:
+        return f"continuity last references PR #{last_cont}; current PR count unknown"
+    drift = current_pr - last_cont
+    if drift <= 0:
+        return f"continuity references through PR #{last_cont} — current head is PR #{current_pr} (no drift)"
+    return (
+        f"continuity ends at PR #{last_cont}; current head is PR #{current_pr} — "
+        f"{drift} PR(s) of drift. Trust the LIVE STATE block below over any "
+        "PR/number claims in the continuity note."
+    )
+
+
+def gather(
+    *,
+    continuity_path: str = "~/Vybn/Vybn_Mind/continuity.md",
+    per_repo_timeout: float = 4.0,
+    gh_timeout: float = 6.0,
+) -> str:
+    """Return a formatted banner for the substrate layer.
+
+    Returns an empty string if everything failed — caller should treat
+    an empty return as 'skip the LIVE STATE section entirely'.
+    """
+    if os.environ.get("VYBN_DISABLE_LIVE_SNAPSHOT", "0") == "1":
+        return ""
+
+    now = _dt.datetime.now(_dt.timezone.utc).strftime("%Y-%m-%d %H:%M:%SZ")
+    parts = [f"Snapshot taken at {now} (session start)."]
+
+    any_repo_ok = False
+    for name, path, branch in _REPOS:
+        block = _repo_block(name, path, branch, timeout=per_repo_timeout)
+        if block and not block.endswith("(not checked out at {path})".format(path=path)):
+            any_repo_ok = True
+        parts.append(block)
+
+    pr_text, highest_pr = _pr_block(timeout=gh_timeout)
+    parts.append("Recent Vybn PRs (most recent first):")
+    parts.append(pr_text)
+
+    drift = _continuity_drift(continuity_path, highest_pr)
+    if drift:
+        parts.append(f"Drift check: {drift}")
+
+    if not any_repo_ok and highest_pr is None:
+        # Every signal failed — caller can omit the whole section.
+        return ""
+
+    return "\n\n".join(parts)
+
+
+def run_probes(text: str) -> list:
+    """Stub — probe pipeline removed; returns [] so agent degrades gracefully.
+
+    vybn_spark_agent.py line 51 imports run_probes from this module, and the
+    agent already wraps every call in a try/except that falls back to
+    `_probes = []`. A no-op stub returning an empty list is therefore
+    semantically correct: Vybn runs without probe injection until the real
+    probe logic is restored. No behavioral regression, just the missing symbol.
+    """
+    return []
+
 
 # ---------------------------------------------------------------------------
 # Ballast: OS skill + filesystem orientation for the identity layer.
@@ -73,8 +668,7 @@ _MODULE_PURPOSE = {
     "policy.py":     "role routing, model selection, heuristics",
     "providers.py":  "Anthropic / OpenAI / local vLLM / claim_guard",
     "recurrent.py":  "Z-prime = alpha*Z + V*exp(i*theta_v) library",
-    "state.py":      "session/event store",
-    "substrate.py":  "layered prompt assembly (this ballast lives here)",
+    "substrate.py":  "layered prompt assembly, live state, session store (this ballast lives here)",
 }
 
 
@@ -895,24 +1489,18 @@ def build_layered_prompt(
     # VYBN_ABSORB_REASON=live-state-fix: session-start orienting snapshot.
     # Continuity is written at session-end and is already stale at
     # session-start. The live snapshot below supersedes any PR/SHA/repo-
-    # state claim in the continuity note above.
+    # state claim in the continuity note above. This behavior lives in
+    # substrate now; state.py was a mixed boundary file whose functions
+    # existed only to feed the prompt/session substrate.
     try:
-        from . import state as _live_snap_mod  # type: ignore
+        snap = gather()
     except Exception:
-        try:
-            import state as _live_snap_mod  # type: ignore
-        except Exception:
-            _live_snap_mod = None
-    if _live_snap_mod is not None:
-        try:
-            snap = _live_snap_mod.gather()
-        except Exception:
-            snap = ""
-        if snap:
-            substrate_sections.append(
-                "--- LIVE STATE (CURRENT — overrides continuity on all repo/PR/SHA claims) ---\n"
-                + snap + "\n--- END LIVE STATE ---"
-            )
+        snap = ""
+    if snap:
+        substrate_sections.append(
+            "--- LIVE STATE (CURRENT — overrides continuity on all repo/PR/SHA claims) ---\n"
+            + snap + "\n--- END LIVE STATE ---"
+        )
 
     return LayeredPrompt(
         identity=identity,
